+++ DiskEmulator.h
@@ -21,12 +21,12 @@ typedef struct
 } DiskArray;
 
 //opens a disk arrary from file
-void openDisk(DiskArray *, UINT);
+void openDisk(DiskArray *, LONG);
 
 //initialize a disk array in memory
 //args: device,
 //      size
-void initDisk(DiskArray *, UINT);
+void initDisk(DiskArray *, LONG);
 
 //destroy the in-memory disk array
 void closeDisk(DiskArray *);
